requirements.txt
torch
torchvision
facenet-pytorch
opencv-python
opencv-python-headless
numpy
pillow
matplotlib
seaborn
scikit-learn
streamlit
deepface

